Rename package to localstack_prometheus
