+++ prometheus/localstack_prometheus/__init__.py
@@ -1,0 +1,1 @@
+name = "localstack_prometheus"
